fix output tag
I had the mistaken idea that it didn't apply normally because my SF settings were messed up. So, it's back to the behaviour of Warlock2 (except output doesn't clear on prompt, which was a bug in Warlock2)

Added a debug window while I was at it. It will just get debug logs from scripts for now.