move text buffering from streams to client
to enable the client to change the destination of text based on content, it needs to collect the entire line in the client before sending it to the stream. Overall this moves some complexity from the stream back to the client. I took this opportunity to fix what looked like problem in the raw output mode.